Move file under pkg/integration folder
